requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
Pillow>=10.0.0
numpy>=1.24.3
pandas>=2.0.3
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
kaggle>=1.5.16
kagglehub>=0.2.0
tqdm>=4.66.1
torch>=2.2.0
torchvision>=0.17.0
